rename GrowSeed.cpp -> GrowPatch.cpp and delete surfacehelpers.hpp
