+++ apps/src/vc_grow_seg_from_segments.cpp
@@ -1,33 +1,24 @@
 #include <nlohmann/json.hpp>
 
 #include "vc/core/util/xtensor_include.hpp"
-#include XTENSORINCLUDE(containers, xarray.hpp)
-#include XTENSORINCLUDE(views, xaxis_slice_iterator.hpp)
 #include XTENSORINCLUDE(io, xio.hpp)
-#include XTENSORINCLUDE(generators, xbuilder.hpp)
 #include XTENSORINCLUDE(views, xview.hpp)
 
 #include "z5/factory.hxx"
 #include "z5/filesystem/handle.hxx"
-#include "z5/filesystem/dataset.hxx"
-#include "z5/common.hxx"
 #include "z5/multiarray/xtensor_access.hxx"
 #include "z5/attributes.hxx"
 
-#include <opencv2/highgui.hpp>
 #include <opencv2/core.hpp>
-#include <opencv2/imgproc.hpp>
 
 #include "vc/core/util/Slicing.hpp"
 #include "vc/core/util/Surface.hpp"
 #include "vc/core/io/PointSetIO.hpp"
 
-#include <unordered_map>
 #include <filesystem>
-#include <omp.h>
 
-#include "../../core/src/SurfaceHelpers.hpp"
 #include "vc/core/types/ChunkedTensor.hpp"
+#include "vc/core/util/DateTime.hpp"
 
 using shape = z5::types::ShapeType;
 using namespace xt::placeholders;
@@ -45,31 +36,6 @@ std::ostream& operator<< (std::ostream& out, const xt::svector<size_t> &v) {
     return out;
 }
 
-class MeasureLife
-{
-public:
-    MeasureLife(std::string msg)
-    {
-        std::cout << msg << std::flush;
-        start = std::chrono::high_resolution_clock::now();
-    }
-    ~MeasureLife()
-    {
-        auto end = std::chrono::high_resolution_clock::now();
-        std::cout << " took " << std::chrono::duration<double>(end-start).count() << " s" << std::endl;
-    }
-private:
-    std::chrono::time_point<std::chrono::high_resolution_clock> start;
-};
-
-
-
-template <typename T, typename I>
-float get_val(I &interp, cv::Vec3d l) {
-    T v;
-    interp.Evaluate(l[2], l[1], l[0], &v);
-    return v;
-}
 
 int main(int argc, char *argv[])
 {
