+++ core/src/Volume.cpp
@@ -155,15 +155,7 @@ void throw_run_path(const fs::path &path, const std::string msg)
     throw std::runtime_error(msg + " for " + path.string());
 }
 
-std::ostream& operator<< (std::ostream& out, const xt::xarray<uint8_t>::shape_type &v) {
-    if ( !v.empty() ) {
-        out << '[';
-        for(auto &v : v)
-            out << v << ",";
-        out << "\b\b]"; // use two ANSI backspace characters '\b' to overwrite final ", "
-    }
-    return out;
-}
+
 
 z5::Dataset *Volume::zarrDataset(int level)
 {
